requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
pkg_resources==0.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.5.4
seaborn==0.11.2
tweepy==4.6.0
python-decouple==3.5
PyYAML==6.0
spacy==3.5.3
wordcloud==1.9.2
Pillow==10.0.1 